Revert "Refaktorert"
