--- blueprints/registrer/__init__.py
@@ -1,3 +1,0 @@
-from .routes import registrer_bp
-
-__all__ = ['registrer_bp']
