--- blueprints/registrer/routes.py
@@ -1,54 +1,0 @@
-from flask import Blueprint, render_template, request, redirect, url_for
-from sql.db_connection import fetch_all_kriser, fetch_all_locations, run_query, get_last_inserted_id, fetch_status_data, update_status, search_statuses
-
-registrer_bp = Blueprint('registrer', __name__)
-
-@registrer_bp.route("/register", methods=["GET", "POST"])
-def register():
-    if request.method == "POST":
-        try:
-            fornavn = request.form.get("fornavn")
-            mellomnavn = request.form.get("mellomnavn")
-            etternavn = request.form.get("etternavn")
-            adresse = request.form.get("adresse")
-            telefonnummer = request.form.get("telefonnummer")
-            status = request.form.get("status")
-            lokasjon_id = request.form.get("lokasjon")
-            parorende_fornavn = request.form.get("parorende_fornavn")
-            parorende_mellomnavn = request.form.get("parorende_mellomnavn")
-            parorende_etternavn = request.form.get("parorende_etternavn")
-            parorende_telefonnummer = request.form.get("parorende_telefonnummer")
-
-            # Insert data into the database
-            query = f"""
-                INSERT INTO Evakuerte (Fornavn, MellomNavn, Etternavn, Adresse, Telefonnummer)
-                VALUES ('{fornavn}', '{mellomnavn}', '{etternavn}', '{adresse}', '{telefonnummer}');
-            """
-            run_query(query)
-
-            # Get the last inserted EvakuertID
-            evakuert_id = get_last_inserted_id()
-
-            # Insert data into the KontaktPerson table
-            query = f"""
-                INSERT INTO KontaktPerson (Fornavn, MellomNavn, Etternavn, Telefonnummer, EvakuertID)
-                VALUES ('{parorende_fornavn}', '{parorende_mellomnavn}', '{parorende_etternavn}', '{parorende_telefonnummer}', {evakuert_id});
-            """
-            run_query(query)
-
-            # Insert data into the Status table
-            query = f"""
-                INSERT INTO Status (Status, Lokasjon, EvakuertID)
-                VALUES ('{status}', '{lokasjon_id}', {evakuert_id});
-            """
-            run_query(query)
-
-            return redirect(url_for("index"))
-        except Exception as e:
-            print(f"An error occurred: {e}")
-            return "An error occurred while processing your request."
-
-    kriser = fetch_all_kriser()  # Fetch all kriser from the database
-    locations = fetch_all_locations()  # Fetch all locations from the database
-    return render_template('register.html', kriser=kriser, locations=locations)
-
